task_3/requirements.txt
# MySQL Server == 8.0.27
pyyaml >= 5.4.1
mysql-connector-python >= 8.0.27
